# backend/docs/swagger.yaml
basePath: /
definitions:
  dto.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  dto.CreateFTPServerRequest:
    properties:
      description:
        type: string
      driver:
        type: string
      host:
        type: string
      max_connections:
        type: integer
      name:
        type: string
      passive_port_end:
        type: integer
      passive_port_start:
        type: integer
      password:
        type: string
      port:
        type: integer
      root_path:
        type: string
      status:
        type: string
      username:
        type: string
    required:
    - host
    - name
    - password
    - root_path
    - username
    type: object
  dto.CreateFTPUserRequest:
    properties:
      download_limit_mbps:
        type: integer
      ftp_server_id:
        type: integer
      home_directory:
        type: string
      password:
        type: string
      remarks:
        type: string
      status:
        type: string
      storage_quota_gb:
        type: integer
      subscription_id:
        type: integer
      upload_limit_mbps:
        type: integer
      username:
        type: string
    required:
    - ftp_server_id
    - home_directory
    - password
    - subscription_id
    - username
    type: object
...
host: localhost:8080
info:
  contact:
    email: ts.communicationmagura@gmail.com
    name: TS Communication
  description: Enterprise ISP Billing & Network Management System
  license:
    name: MIT
  title: TS Cloud ISP Billing API
  version: "1.0"
paths:
...
  /api/v1/ftp-servers:
    get:
      description: Get all FTP servers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: List FTP Servers
      tags:
      - FTP Server
    post:
      consumes:
      - application/json
      description: Create a new FTP server
      parameters:
      - description: FTP Server
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateFTPServerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Create FTP Server
      tags:
      - FTP Server
  /api/v1/ftp-servers/{id}:
    delete:
      description: Delete FTP server by ID
      parameters:
      - description: FTP Server ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete FTP Server
      tags:
      - FTP Server
    get:
      description: Get FTP server by ID
      parameters:
      - description: FTP Server ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Get FTP Server
      tags:
      - FTP Server
    put:
      consumes:
      - application/json
      description: Update FTP server information
      parameters:
      - description: FTP Server ID
        in: path
        name: id
        required: true
        type: integer
      - description: FTP Server
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateFTPServerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Update FTP Server
      tags:
      - FTP Server
  /api/v1/ftp-users:
    get:
      description: Get all FTP users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: List FTP Users
      tags:
      - FTP User
    post:
      consumes:
      - application/json
      description: Create a new FTP user
      parameters:
      - description: FTP User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateFTPUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Create FTP User
      tags:
      - FTP User
  /api/v1/ftp-users/{id}:
    delete:
      description: Delete FTP user by ID
      parameters:
      - description: FTP User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete FTP User
      tags:
      - FTP User
    get:
      description: Get FTP user by ID
      parameters:
      - description: FTP User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Get FTP User
      tags:
      - FTP User
    put:
      consumes:
      - application/json
      description: Update FTP user information
      parameters:
      - description: FTP User ID
        in: path
        name: id
        required: true
        type: integer
      - description: FTP User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateFTPUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Update FTP User
      tags:
      - FTP User
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"